Added state module
This module will be expanded in future once we understand how continuous
and infinite state representations work. For now, this module contains
an AttrStateSpace and AttrState set of classes, to make work with state
indexes and lookups much easier when doing things like making reward,
value, transition functions, and when evaluating policies.