Adds trained dqn agent
This change adds a trained dqn agent to the project. It has been trained
on 5,400,000 iterations of the game blackjack. Current issues with it:
1. Avg. winning multiplier per hand is too high. After evaluating model
on 500,000 hands of blackjack (shuffling every new hand) the avg. winning
multiplier per hand is 0.0738483. We would expect this number to be slightly
negative indicating that the player is losing money over time on a constant bet
2. Average reward per episode is also slightly too high. We would expect
the avg. reward per episode to be a slightly negative number indicating player
loses more over time.