Retrains DQN
Fixed a bug where the player had access to the dealer's whole hand value in its state
Haven't been able to get the DQN to converge to the optimal solution for blackjack
strategy. It seems like the NN is getting stuck in local optima or that because
of the randomness of the game of blackjack it is unable to explore enough of the
state space to be able to figure out optimal strategy in each state.