Adds trained dqn agent (#5)
* Adds trained dqn agent

This change adds a trained dqn agent to the project. It has been trained
on 5,400,000 iterations of the game blackjack. Current issues with it:
1. Avg. winning multiplier per hand is too high. After evaluating model
on 500,000 hands of blackjack (shuffling every new hand) the avg. winning
multiplier per hand is 0.0738483. We would expect this number to be slightly
negative indicating that the player is losing money over time on a constant bet
2. Average reward per episode is also slightly too high. We would expect
the avg. reward per episode to be a slightly negative number indicating player
loses more over time.

* Cleans up evaluation and training scripts

This change cleans up the evaluation and training scripts

* Adds trained dqn agent

This change adds a trained dqn agent to the project. It has been trained
on 5,400,000 iterations of the game blackjack. Current issue with it is that
the DQN has not achieved the optimal solution.

* Cleans up evaluation and training scripts

This change cleans up the evaluation and training scripts

* Retrains DQN

Fixed a bug where the player had access to the dealer's whole hand value in its state
Haven't been able to get the DQN to converge to the optimal solution for blackjack
strategy. It seems like the NN is getting stuck in local optima or that because
of the randomness of the game of blackjack it is unable to explore enough of the
state space to be able to figure out optimal strategy in each state.

* Adds trained dqn agent

This change adds a trained dqn agent to the project. It has been trained
on 5,400,000 iterations of the game blackjack. Current issue with it is that
the DQN has not achieved the optimal solution.

* Cleans up evaluation and training scripts

This change cleans up the evaluation and training scripts

* Cleans up evaluation and training scripts

This change cleans up the evaluation and training scripts

* Fixes another bug where the dealer card was shown to the dealer incorrectly.

* Added error if no blackjack and index still out of bounds

Co-authored-by: rsmith49 <ryans_cais@yahoo.com>